pyproject.toml
[project]
name = "mcp-daytona-server"
version = "0.1.0"
description = "MCP Server using Daytona SDK for LLM code execution"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Nikola Balic", email = "niko@daytona.io" }]
dependencies = [
    "mcp[cli]",
    "daytona-sdk",
    "httpx",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src"]

[tool.hatch.build.targets.wheel]
only-packages = true

[tool.pyright]
include = ["src"]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py310"

[project.scripts]
mcp-daytona = "src.main:main"
